requirements.txt
fastapi
uvicorn
httpx
python-dotenv
itsdangerous
mcp
pydantic